Extract sandbox garden functions into gardenSandboxRepo
